Create message for the git diff:
```
Add message box and text box creation in Pamyu initHtmlElements
```

This change adds the creation of message box and text box in the `initHtmlElements` method of the `Pamyu` class. The message box is created as a `div` element with the id "message-box" and a background image is set using the asset manager. The text box is created as a `div` element with the id "text-box". Both elements are appended to the container, and a new `MessageManager` instance is created.