feat: Add new method to set expression pattern in IAssetManager
Add a new method `setExpressionPattern` to the `IAssetManager` interface. This method takes a `pattern` string as input and allows you to set a pattern for expressions. The pattern can include the following placeholders: `{character}` for the character name, `{side}` for the left or right side, and `{expression}` for the expression name. This method is used to customize expressions in the asset manager.