src-tauri: Add list_res command to handle listing resources
This commit adds a new command `list_res` to handle listing resources in a specified directory. The `list_res` command uses the `fs::read_dir` function to read the contents of the directory, and then iterates over the entries, collecting the file names into a vector. If any invalid UTF-8 sequence is encountered in a file name, the command returns an error with the message "Invalid UTF-8 sequence in file name". If there is any other error reading the directory, the command returns the corresponding error message.

The `list_res` command is registered as an invoke handler in the `main` function of the Tauri application.

Note: This commit also includes modifications to the `Cargo.toml` file to add the "path-all" feature to the `tauri` dependency.