Add action ordering and related functionalities to scenesStore
This commit introduces the order in which actions are performed in scenesStore.ts. An '_order' property has been added to the action object to manage this sequence. Functions to get an action by order, to update an action's order and to invert the order of two actions have also been implemented. Now, actions can be sorted and retrieved by their order. This change will facilitate a more structured management of actions and their execution sequence.