Update elements in the action component and add order
This commit updates 'Timeline.vue', 'Action.vue', and 'scene.ts' components and adds a new 'ActionBtn.vue' component to the scene directory. The main update is the addition of an '_order' property to the 'BaseElement' in scene.ts, which is used to manage the order of actions. This is particularly used to manage drag and drop functionality in 'Timeline.vue'.

Moreover, some changes for code enhancement and bug fix were made in 'Timeline.vue' and 'Action.vue'. The paths and import statements were updated. We replaced the 'Icon' component with the new 'ActionBtn' component in 'Action.vue'. Also a 'deleteAction', 'orderUp', and 'orderDown' methods were added and others were improved.

The 'ActionBtn.vue' component was added to replace the delete button functionality in the 'Action.vue' component with a more generic button that can be clicked to trigger any action. This was implemented making usability improvements for brake and improve maintainability.

Lastly, a minor update in 'InputText.vue' to make the 'label' property optional.

The changes represent significant improvements in the code maintainability, structure, and user interaction with the action elements in the application.