Update dependencies and improve Actions functionality
Reviewing the code changes, there are two main things addressed in this commit. Firstly, the functionality of Actions was improved to allow for their deletion, an important addition for better user experience and system's flexibility. This update is now reflected in the CHANGELOG.md. Secondly, several of our dependencies in package.json have been updated to their latest versions; this includes eslint-plugin-vue, sass, vue-i18n, and vue-tsc. Routinely updating our dependencies helps keep the project secure and up-to-date, ensuring we are using the most recent features, improvements, and security patches.