Improve Action component and include transition groups
In this commit, the Vue's built-in `<TransitionGroup>` tag was included in the Timeline component, specifically wrapping the Action instances. This use of transition groups is for the purpose of providing animations when the list of actions is reordered or when an action is added or deleted. This improves the user interface by making it more interactive and intuitive. Additionally, a `key` attribute was added to the <TimelineAction> component as a good Vue practice.