Add 'message' action functionality and translations
Adding a 'message' functionality was essential to provide users with new interaction type. Now users can create a message action and specify a character in scenes. The 'message' action can be used without assigning it to a character as well. 'message' option was also added to the action type switch statement in Timeline.vue. Translations for 'message' and 'noCharacter' were also added for multiple languages.