Refactored utils and added 'message' action type
The 'message' action type was added to the list of actions in actionsElements.ts while refining the code structure. In utils/tools.ts, the argument for getIndent function is now given a default value of 1. In utils/pamyu.ts, 'message' was added to the import statement and its corresponding functionality was implemented in the getMessageActionLines function. This is to enhance the interaction capabilities of users within scenes.