Add delete_file function and file import capabilities
This commit adds a delete_file function to the files module in Tauri, and also updates the main.rs file to import this new functionality. The delete_file function is to allow the application to remove files, it checks if the provided path is a file and if it exists, then removes it. The main.rs changes basically include the delete_file function as part of the list of functions accessible to the app and also includes an additional module for file import. This commit also updates Cargo.toml and Cargo.lock files to include 'walkdir', a new dependency for directory tracking, which is crucial for file import functionality. Lastly, a new file named import.rs has been added. This is a dedicated module to handle all Ren'Py project files retrieval operation. This further enhances the management and tracking of files in Ren'Py projects.