Refactor scripts scanning test and add script processing to ScanManager
Refactored, expanded and relocated the script scanning tests to cover more cases. Created ScanManager class for handling the parsing of scripts into blocks, which is vital for subsequent script analysis and transformation. Made modifications in import.ts to improve file management. This not only helps in comprehensively testing the script scanning process but also provides clearer guidelines for future code maintainability and extension.